badrock/Cypher/requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
google-generativeai>=0.8.0

